Rename CefWindowInfoWrapper.h to WindowInfo.h as it's not a public class that can be instantiated for use with ShowDevTools
