+++ CefSharp.Core/CefSharp.Core.vcxproj
@@ -267,7 +267,7 @@
     <ClInclude Include="Internals\CefSharpBrowserWrapper.h" />
     <ClInclude Include="Internals\CefTaskScheduler.h" />
     <ClInclude Include="Internals\CefTaskWrapper.h" />
-    <ClInclude Include="Internals\CefWindowInfoWrapper.h" />
+    <ClInclude Include="WindowInfo.h" />
     <ClInclude Include="Internals\CompletionHandler.h" />
     <ClInclude Include="Internals\ClientAdapter.h" />
     <ClInclude Include="Internals\CookieManager.h" />
